modificación y refactorización de Productos
