+++ multi-stock-sync/src/Sync/Views/Productos/Router/RouterProducto.tsx
@@ -1,6 +1,6 @@
 import { Route, Routes } from "react-router-dom";
 import GestionProducto from "../Views/GestionProducto";
-import CrearProducto from "../Views/CrearProducto";
+import CrearProducto from "../Views/components/CrearProducto"; // ✅ corregido
 import CargaMasiva from "../Views/CargaMasiva"; // ✅ corregido
 
 const RouterProducto = () => {
